+++ var/starter_kits/italiagov/webpack.common.js
@@ -1,6 +1,8 @@
+const fs = require('fs')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const CopyWebpackPlugin = require('copy-webpack-plugin')
 const SpriteLoaderPlugin = require('svg-sprite-loader/plugin')
+const rimraf = require('rimraf')
 const check = require('./webpack.check')
 
 const paths = require('./webpack.paths')
@@ -9,7 +11,9 @@ module.exports = {
   // Entry
   entry: {
     "bootstrap-italia": [paths.src + '/js/index.js', paths.src + '/scss/theme.scss'],
-    //"bootstrap-italia-comuni": [paths.src + '/js/index.js', paths.src + '/scss/theme-comuni.scss']
+    //"bootstrap-italia-comuni": [paths.src + '/js/index.js', paths.src + '/scss/theme-comuni.scss'],
+    //"ckeditor5": paths.src + '/scss/ckeditor5.scss',
+    //"ckeditor5-comuni": paths.src + '/scss/ckeditor5-comuni.scss',
   },
 
   // Output
@@ -60,7 +64,7 @@ module.exports = {
       plainSprite: true
     }),
     new CopyWebpackPlugin({
-      patterns:[
+      patterns: [
         {
           from: paths.modules + '/bootstrap-italia/src/assets/',
           to: paths.build + '/assets/'
@@ -75,5 +79,19 @@ module.exports = {
         }
       ]
     }),
+    {
+      apply: (compiler) => {
+        compiler.hooks.afterEmit.tap('AfterEmitPlugin', (compilation) => {
+          const ckeditorJsFile = compiler.options.output.path + '/js/ckeditor5.min.js';
+          const ckeditorComuniJsFile = compiler.options.output.path + '/js/ckeditor5-comuni.min.js';
+          if (fs.existsSync(ckeditorJsFile)) {
+            rimraf.sync(ckeditorJsFile);
+          }
+          if (fs.existsSync(ckeditorComuniJsFile)) {
+            rimraf.sync(ckeditorComuniJsFile);
+          }
+        });
+      },
+    }
   ],
 };